Reorganise HAL with the HTTP adapters
